--- src/main/java/com/teragrep/Request.java
@@ -1,10 +1,0 @@
-package com.teragrep;
-
-import org.apache.http.HttpResponse;
-
-import java.io.IOException;
-
-public interface Request {
-
-    HttpResponse doRequest() throws IOException;
-}
